Added requested detection of Operating System and Version number.
I Added a LOT of OS's and their Version numbers. I got them from a website(http://user-agent-string.info/list-of-ua/os). They are all calculated out of the UserAgent. This gives the downside that:
Nokia version detection is impossible. They only add Nokia to the UserAgent. Nothing more.
Detection if tablet or mobile running on Android is mobile or tablet is impossable. This can only be fixed if all the possible android tablet/mobile combination useragents are added in this file and walked through.

The version numbers are updated until the day of now ofcourse. If there is, lets say iOS9, then there will be a fall back on iOS. Take notice that the order in which the OS's are placed is crucial. For instance, the consoles must be checked before windows, becuase the xBox also uses Windows Nt 6.1, but trident is added to it.

If you notice performance issues on the script, then you can ofcourse remove the other part of operating systems, because some of them date back until 1994.

Good luck, hope it helped.